feat(#65): 지도 실시간 위치 표시시 프로필 사진 띄우기

+++ Where_Are_You/Data/Models/Responses/Schedule/GetScheduleResponse.swift
@@ -10,7 +10,7 @@ import Foundation
 struct GetScheduleResponse: Codable {
     let title, startTime, endTime, location: String
     let streetName: String
-    let x, y: Int
+    let x, y: Double
     let color, memo: String
     let allDay: Bool
     let memberInfos: [MemberInfos]
